Year 3
The gap is growing...

Ædam 粘土人 sees nothing wrong with keep_grabbing.py. 

An n-counter with an MIT Firewall, a Profess.ed⁺ Hacker Haven, is no different from my n-counters tunneling thru the Great Firewall of 中国 with MACintosh Address Change and The Wild Onion Router.

JSTOR, et. al., are the Trees of Life, and the Trees of Knowledge of Good & Evil. These Trees, with Fruits Once Tasted, must be Replanted and Reclimbed. The larger the Harvests, the finer Our Lives, given that We're willing to Undergo the Work that Comes with Knowledge, not just Accumulate Knowledge without Bearing its ∱TEAM Fruits. The alternative is life-and-death; dropping prevention, falling to rescue, and then failing to triage. Having built these Trees with the Civic and Civilizational Tithe Work of Our Own Hhands, We are Entitled to the Outcomes, the Fruits of Our Labors, which Promised an Education⁺ in an Educate.ed⁺ Democracy⁵. The last moments of formatting the Tithe do not transform the Tithe Work into the Private Domain.  The last moments of formatting do not and should not impede the Spread of Wisdom in Democracy⁵. 

https://archive.org/stream/GuerillaOpenAccessManifesto/Goamjuly2008_djvu.txt

Time after time, in studying and following Policy Proposals for Shuttle Runs to the U|S⁺ Senate, in efforts to keep up with the Wave of Reservation Service i might have had as a net present impact physician, pay barriers on Publicly Fund.ed⁺ Research interrupted my Citizenship Work. The Access Grate for Hhippocratic Hhands should not have to include active tuition payments to an Élite Research Institution.. Good Will Hunting, reading from hhome, should be able to outclip the insight of a Research Faculty Member, tracing a Thesis that the Disciplines missed. 

There are Limits to Sharing, just as there are Limits to Copyright. When the Limits to Copyright are Systematically Undermined, the other Balance of the Equation does not need hold its End of the Bargain.  

Y⁵ou can add Ædam's Name to the List.